+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 /* lib hooks */
 export { default as useOutsideClick } from './lib/useOutsideClick';
+export { default as useStepValidation } from './lib/useStepValidation';
 
 /* quries hooks */
 export { default as useGetExample } from './queries/useGetExample';
